Change poll time if no events occur.
The main problem we encounter here is that we want to independently wait
for two interrupts which both can't be shared with another thread, as
they are non-send depend on another (WlSurface <=> WlOutput <=>
EventQueue), we would like to merge this to a single interrupt to which
we can just idly wait. Though due to structure this is made difficult.